Register data-bus reads and writes, and allow variable data-bus latency.
The "rreq" (read-request) strobe is now accompanied with a "rack"
(read-acknowledge) strobe that accompanies the data return. The
processor pipeline is stalled until each rreq is answered with a rack.

It is not as simple to implement variable instruction-read latency,
because the instruction bus is pipelined and multiple reads overlap.

Note that the size of minimax just jumped significantly (it's in the
README). Because the CPU does not reach timing closure at 100 MHz, I'm
willing to spend registers and LUTs to break up the critical path. I'm
also expecting to put the core on a diet at some point in the future.