Adds TINY mode for microcode < 0x400 = 1024 bytes.
At commit time, this microcode build is 0x3b2 = 946 bytes large. The
goal for sticking under 1024 bytes is to tuck it into a quarter of a 4kB
block RAM (RAMB36) and give the remaining program space back to the CPU.

An alternative would be to move the trap vector somewhere else (e.g.
0xffe) so that we don't have to statically divide BRAM between microcode
and user code.